requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
pydantic[email]
sqlalchemy
psycopg2
dotenv
passlib[bcrypt]
python-jose
